chore: move screen-specific components to @berty/screens
