+++ js/packages/components/shared-components/Button.tsx
@@ -5,7 +5,7 @@ import { useAppDimensions } from '@berty/contexts/app-dimensions.context'
 import { useStyles } from '@berty/contexts/styles'
 import { useThemeColor } from '@berty/store/hooks'
 
-import { UnifiedText } from '../shared-components/UnifiedText'
+import { UnifiedText } from './UnifiedText'
 
 const Button: React.FC<
 	{
